Add the license and notice for ConcurrentMapBackedCache.scala
The source code file called ConcurrentMapBackedCache.scala uses the source
code of Spray.

Closes: apache/incubator-openwhisk-release#152
