+++ common/scala/src/main/scala/whisk/core/database/MultipleReadersSingleWriterCache.scala
@@ -15,21 +15,14 @@
  * limitations under the License.
  */
 
-/*
- * Cache base implementation:
- * Copyright (C) 2017 Lightbend Inc. <http://www.lightbend.com/>
- */
-
 package whisk.core.database
 
-import java.util.concurrent.{ConcurrentMap, TimeUnit}
+import java.util.concurrent.TimeUnit
 import java.util.concurrent.atomic.AtomicReference
 
 import scala.concurrent.{ExecutionContext, Future, Promise}
-import scala.language.implicitConversions
 import scala.util.Failure
 import scala.util.Success
-import scala.util.control.NonFatal
 import com.github.benmanes.caffeine.cache.Caffeine
 import whisk.common.Logging
 import whisk.common.LoggingMarkers
@@ -473,52 +466,4 @@ trait MultipleReadersSingleWriterCache[W, Winfo] {
           .build()
           .asMap())
   }
-}
-
-/**
- * A thread-safe implementation of [[spray.caching.cache]] backed by a plain
- * [[java.util.concurrent.ConcurrentMap]].
- *
- * The implementation is entirely copied from Spray's [[spray.caching.Cache]] and
- * [[spray.caching.SimpleLruCache]] respectively, the only difference being the store type.
- * Implementation otherwise is identical.
- */
-private class ConcurrentMapBackedCache[V](store: ConcurrentMap[Any, Future[V]]) {
-  val cache = this
-
-  def apply(key: Any) = new Keyed(key)
-
-  class Keyed(key: Any) {
-    def apply(magnet: => ValueMagnet[V])(implicit ec: ExecutionContext): Future[V] =
-      cache.apply(
-        key,
-        () =>
-          try magnet.future
-          catch { case NonFatal(e) => Future.failed(e) })
-  }
-
-  def apply(key: Any, genValue: () => Future[V])(implicit ec: ExecutionContext): Future[V] = {
-    store.computeIfAbsent(
-      key,
-      new java.util.function.Function[Any, Future[V]]() {
-        override def apply(key: Any): Future[V] = {
-          val future = genValue()
-          future.onComplete { value =>
-            // in case of exceptions we remove the cache entry (i.e. try again later)
-            if (value.isFailure) store.remove(key, future)
-          }
-          future
-        }
-      })
-  }
-
-  def remove(key: Any) = Option(store.remove(key))
-
-  def size = store.size
-}
-
-class ValueMagnet[V](val future: Future[V])
-object ValueMagnet {
-  implicit def fromAny[V](block: V): ValueMagnet[V] = fromFuture(Future.successful(block))
-  implicit def fromFuture[V](future: Future[V]): ValueMagnet[V] = new ValueMagnet(future)
 }
